requirements.txt
discord
googletrans
python-decouple